% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-utils.R
\name{expand_dates_from_regex}
\alias{expand_dates_from_regex}
\title{Function to expand dates from a regex}
\usage{
expand_dates_from_regex(date_regex, ver = c(1, 2))
}
\arguments{
\item{date_regex}{A regular expression to match dates in the format yyyymmdd.}

\item{ver}{The version of the data to use. Defaults to "v1". Can be "v1" or "v2".}
}
\value{
A character vector of dates matching the regex.
}
\description{
This function generates a sequence of dates from a regular expression pattern.
based on the provided regular expression.
}
\keyword{internal}
