% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_data_dir}
\alias{get_data_dir}
\title{Get the data directory}
\usage{
get_data_dir()
}
\value{
The data directory.
}
\description{
This function retrieves the data directory from the environment variable SPANISH_OD_DATA_DIR.
If the environment variable is not set, it returns the temporary directory.
}
\keyword{internal}
