% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_metadata}
\alias{get_metadata}
\title{Get the data dictionary}
\usage{
get_metadata(data_dir = get_data_dir())
}
\arguments{
\item{data_dir}{The directory where the data is stored. Defaults to the value returned by \code{get_data_dir()}.}
}
\value{
The data dictionary.
}
\description{
This function retrieves the data dictionary for the specified data directory.
}
\examples{
# Get the data dictionary for the default data directory
metadata = get_metadata()
names(metadata)
head(metadata)
}
