% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-params.R
\name{global_quiet_param}
\alias{global_quiet_param}
\title{Global Quiet Parameter}
\usage{
global_quiet_param(quiet = FALSE)
}
\arguments{
\item{quiet}{A \code{logical} value indicating whether to suppress messages. Default is \code{FALSE}.}
}
\value{
Nothing. This function is just a placeholder for global quiet parameter.
}
\description{
Documentation for the \code{quiet} parameter, used globally.
}
\keyword{internal}
