% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-utils.R
\name{process_date_arguments}
\alias{process_date_arguments}
\title{Process multiple date arguments}
\usage{
process_date_arguments(
  date_range = NULL,
  dates_list = NULL,
  date_regex = NULL,
  ver = c(1, 2)
)
}
\arguments{
\item{date_range}{A vector of dates in ISO format (yyyy-mm-dd) or yyyymmdd format.}

\item{dates_list}{A vector of dates in ISO format (yyyy-mm-dd) or yyyymmdd format.}

\item{date_regex}{A regular expression to match dates in the format yyyymmdd.}
}
\value{
A Dates vector of dates.
}
\description{
This function processes the date arguments provided to various functions in the package.
It checks if more than one date argument is provided and returns the appropriate dates as a Dates vector.
The function ensures that the dates are in ISO format (yyyy-mm-dd) or yyyymmdd format.
}
\keyword{internal}
