% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-v1-data.R
\name{spo_load_zones_v1}
\alias{spo_load_zones_v1}
\title{Loads the fixed zones for v1 data}
\usage{
spo_load_zones_v1(data_dir = get_data_dir(), type = "distritos")
}
\arguments{
\item{data_dir}{The directory where the data is stored.}

\item{type}{The type of zones data to retrieve ("distritos" or "municipios").}
}
\value{
A spatial object of type \code{sf} containing the fixed zones data.
\describe{
\item{id}{\code{character}. The identifier of the zone.}
}
}
\description{
This function loads the fixed zones data for the specified data directory.
The geometry is cheked for validity and fixed if necessary. The 'ID' column is renamed to 'id'.
}
\examples{
if (FALSE) {
  zones <- spo_load_zones_v1()
}
}
