% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available-data-s3.R
\name{spod_available_data_s3}
\alias{spod_available_data_s3}
\title{Get available data list from Amazon S3 storage}
\usage{
spod_available_data_s3(ver = c(1, 2), s3_force_update = FALSE)
}
\arguments{
\item{ver}{Integer. Can be 1 or 2. The version of the data to use. v1 spans 2020-2021, v2 covers 2022 and onwards.}

\item{s3_force_update}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Logical. If \code{TRUE}, force update of the available data list from Amazon S3. If \code{FALSE}, only update the available data list if it is older than 1 day.}
}
\value{
A tibble with links, release dates of files in the data, dates of data coverage, local paths to files, and the download status.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Get a table with links to available data files for the specified data version from Amazon S3 storage.
}
