% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_v1_data.R
\name{spod_available_data_v1}
\alias{spod_available_data_v1}
\title{Get the available v1 data list}
\usage{
spod_available_data_v1(
  data_dir = spod_get_data_dir(),
  check_local_files = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{data_dir}{The directory where the data is stored. Defaults to the value returned by \code{spod_get_data_dir()}.}
}
\value{
A tibble with links, release dates of files in the data, dates of data coverage, local paths to files, and the download status.
\describe{
\item{target_url}{\code{character}. The URL link to the data file.}
\item{pub_ts}{\code{POSIXct}. The timestamp of when the file was published.}
\item{file_extension}{\code{character}. The file extension of the data file (e.g., 'tar', 'gz').}
\item{data_ym}{\code{Date}. The year and month of the data coverage, if available.}
\item{data_ymd}{\code{Date}. The specific date of the data coverage, if available.}
\item{local_path}{\code{character}. The local file path where the data is stored.}
\item{downloaded}{\code{logical}. Indicator of whether the data file has been downloaded locally.}
}
}
\description{
This function provides a table of the available data list of MITMA v1 (2020-2021), both remote and local.
}
\examples{
# Get the available v1 data list for the default data directory
if (FALSE) {
  metadata <- spod_available_data_v1()
  names(metadata)
  head(metadata)
}
}
