% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{spod_available_data_v2}
\alias{spod_available_data_v2}
\title{Get the data dictionary}
\usage{
spod_available_data_v2(
  data_dir = spod_get_data_dir(),
  check_local_files = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{data_dir}{The directory where the data is stored. Defaults to the value returned by \code{spod_get_data_dir()}.}

\item{check_local_files}{Whether to check if the local files exist. Defaults to \code{FALSE}.}

\item{quiet}{A logical value indicating whether to suppress messages. Default is \code{FALSE}.}
}
\value{
The data dictionary.
}
\description{
This function retrieves the data dictionary for the specified data directory.
}
\examples{
# Get the data dictionary for the default data directory
if (FALSE) {
  metadata <- spod_available_data_v2()
  names(metadata)
  head(metadata)
}
}
