% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-utils.R
\name{spod_available_ram}
\alias{spod_available_ram}
\title{Get available RAM}
\usage{
spod_available_ram()
}
\value{
A \code{numeric} amount of available RAM in GB.
}
\description{
Get available RAM
}
\keyword{internal}
