% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_v1_data.R
\name{spod_clean_zones_v1}
\alias{spod_clean_zones_v1}
\title{Fixes common issues in the zones data and cleans up variable names}
\usage{
spod_clean_zones_v1(zones_path)
}
\arguments{
\item{zones_path}{The path to the zones spatial data file.}
}
\value{
A spatial object of class \code{sf}.
}
\description{
This function fixes any invalid geometries in the zones data and renames the "ID" column to "id".
}
\keyword{internal}
