% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-zones.R
\name{spod_clean_zones_v2}
\alias{spod_clean_zones_v2}
\title{Fixes common issues in the zones data and cleans up variable names}
\usage{
spod_clean_zones_v2(zones_path)
}
\arguments{
\item{zones_path}{The path to the zones spatial data file.}
}
\value{
A spatial object containing the cleaned zones data.
}
\description{
This function fixes any invalid geometries in the zones data and renames the "ID" column to "id". It also attacches the population counts and zone names provided in the csv files supplied by the original data provider.
}
\keyword{internal}
