% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{spod_codebook}
\alias{spod_codebook}
\title{View codebooks for v1 and v2 open mobility data}
\usage{
spod_codebook(ver = 1)
}
\arguments{
\item{ver}{An \code{integer} or \code{numeric} value. The version of the data. Defaults to 1. Can be 1 for v1 (2020-2021) data and 2 for v2 (2022 onwards) data.}
}
\value{
Nothing, calls relevant vignette.
}
\description{
Opens relevant vignette.
}
