% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-files.R
\name{spod_compute_s3_etag}
\alias{spod_compute_s3_etag}
\title{Compute ETag for a file}
\usage{
spod_compute_s3_etag(file_path, part_size = 8 * 1024^2)
}
\arguments{
\item{file_path}{Character. The path to the file.}

\item{part_size}{Numeric. The size of each part in bytes. Do not change, as this is a default for S3 Etag.}
}
\value{
Character. The ETag for the file.
}
\description{
Compute ETag for a file
}
\keyword{internal}
