% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-utils.R
\name{spod_convert_dates_to_ranges}
\alias{spod_convert_dates_to_ranges}
\title{Convert dates to ranges}
\usage{
spod_convert_dates_to_ranges(dates)
}
\arguments{
\item{dates}{A \code{character} vector of dates.}
}
\value{
A \code{character} vector of date ranges.
}
\description{
This internal helper function reduces a vector of dates to a vector of date ranges to shorten the warning and error messages that mention the valid date ranges.
}
\keyword{internal}
