% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-v1-data.R
\name{spod_download_tables}
\alias{spod_download_tables}
\title{Download the data files of specified type, zones, dates and data version}
\usage{
spod_download_tables(
  type = c("od", "origin-destination", "os", "overnight_stays", "tpp",
    "trips_per_person"),
  zones = c("districts", "dist", "distr", "municipalities", "muni", "municip"),
  date_range = NULL,
  dates_list = NULL,
  date_regex = NULL,
  ver = 1,
  data_dir = spod_get_data_dir()
)
}
\arguments{
\item{type}{The type of data to download. Can be "origin-destination" (or ust "od"), or "trips_per_person" (or just "tpp") for v1 data. For v2 data "overnight_stays" (or just "os") is also available. More data types to be supported in the future. See respective codebooks for more information. \link{ADD CODEBOOKS!}}

\item{zones}{The zones for which to download the data. Can be "districts" (or "dist", "distr") or "municipalities" (or "muni", "municip") for v1 data. Additionaly, these can be "urban_areas" (GAUs) for v2 data.}

\item{date_range}{A character vector of dates in ISO format (YYYY-MM-DD) to download the data for.}

\item{dates_list}{A character vector of dates in ISO format (YYYY-MM-DD) to download the data for. Defaults to NULL.}

\item{date_regex}{A regular expression to match the dates of the data to download. Defaults to NULL.}

\item{ver}{The version of the data to use. Defaults to 1. Can be 1 or 2.}

\item{data_dir}{The directory where the data is stored. Defaults to the value returned by \code{spod_get_data_dir()} which returns the value of the environment variable \code{SPANISH_OD_DATA_DIR} or a temporary directory if the variable is not set.}
}
\description{
This function downloads the data files of the specified type, zones, dates and data version.
}
