% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-helpers.R
\name{spod_duckdb_create_province_enum}
\alias{spod_duckdb_create_province_enum}
\title{Create province names ENUM in a duckdb connection}
\usage{
spod_duckdb_create_province_enum(con)
}
\arguments{
\item{con}{A \code{duckdb} connection.}
}
\value{
A \code{duckdb} connection with \code{INE_PROV_NAME_ENUM} and \code{INE_PROV_CODE_ENUM} created.
}
\description{
Create province names ENUM in a duckdb connection
}
\keyword{internal}
