% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-helpers.R
\name{spod_duckdb_enable_progress_bar}
\alias{spod_duckdb_enable_progress_bar}
\title{Enable progress bar for a \code{DuckDB} connection}
\usage{
spod_duckdb_enable_progress_bar(con, progress_bar_delay_ms = 1000)
}
\arguments{
\item{con}{A \code{duckdb} connection}

\item{progress_bar_delay_ms}{The delay in milliseconds between progress bar updates. Defaults to 1000.}
}
\value{
A \code{duckdb} connection.
}
\description{
Enable progress bar for a \code{DuckDB} connection
}
\keyword{internal}
