% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_helpers.R
\name{spod_duckdb_trips_per_person}
\alias{spod_duckdb_trips_per_person}
\title{Create a duckdb trips per person table}
\usage{
spod_duckdb_trips_per_person(
  con = DBI::dbConnect(duckdb::duckdb(), dbdir = ":memory:", read_only = FALSE),
  zones = c("districts", "dist", "distr", "distritos", "municipalities", "muni",
    "municip", "municipios", "lau", "large_urban_areas", "gau", "grandes_areas_urbanas"),
  ver = NULL,
  data_dir = spod_get_data_dir()
)
}
\arguments{
\item{con}{A duckdb connection object. If not specified, a new in-memory connection will be created.}

\item{zones}{The zones for which to download the data. Can be \code{"districts"} (or \code{"dist"}, \code{"distr"}, or the original Spanish \code{"distritos"}) or \code{"municipalities"} (or \code{"muni"}, \code{"municip"}, or the original Spanish \code{"municipios"}) for both data versions. Additionaly, these can be \code{"large_urban_areas"} (or \code{"lau"}, or the original Spanish \code{"grandes_areas_urbanas"}, or \code{"gau"}) for v2 data (2022 onwards).}

\item{ver}{Integer. Can be 1 or 2. The version of the data to use. v1 spans 2020-2021, v2 covers 2022 and onwards.}

\item{data_dir}{The directory where the data is stored. Defaults to the value returned by \code{spod_get_data_dir()}.}
}
\value{
A duckdb connection with 2 views.
}
\description{
This function creates a duckdb connection to the trips per person data stored in a folder of CSV.gz files.
}
\keyword{internal}
