% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{spod_expand_dates_from_regex}
\alias{spod_expand_dates_from_regex}
\title{Function to expand dates from a regex}
\usage{
spod_expand_dates_from_regex(date_regex)
}
\arguments{
\item{date_regex}{A regular expression to match dates in the format yyyymmdd.}
}
\value{
A character vector of dates matching the regex.
}
\description{
This function generates a sequence of dates from a regular expression pattern.
based on the provided regular expression.
}
\keyword{internal}
