% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick-get.R
\name{spod_fetch_municipalities_json_memoised}
\alias{spod_fetch_municipalities_json_memoised}
\title{Cache the municipalities geometries from the mapas-movilidad website}
\usage{
spod_fetch_municipalities_json_memoised()
}
\description{
Cache the municipalities geometries from the mapas-movilidad website
}
\keyword{internal}
