% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-tools.R
\name{spod_files_sizes}
\alias{spod_files_sizes}
\title{Get files sizes for remote files of v1 and v2 data and save them into a csv.gz file in the inst/extdata folder.}
\usage{
spod_files_sizes(ver = 2)
}
\arguments{
\item{ver}{The version of the data (1 or 2). Can be both. Defaults to 2, as v1 data is not being updated since 2021.}
}
\value{
Nothing. Only saves a csv.gz file with up to date file sizes in the inst/extdata folder.
}
\description{
Get files sizes for remote files of v1 and v2 data and save them into a csv.gz file in the inst/extdata folder.
}
\keyword{internal}
