% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{spod_get_data_dir}
\alias{spod_get_data_dir}
\title{Get the data directory}
\usage{
spod_get_data_dir(quiet = FALSE)
}
\arguments{
\item{quiet}{A logical value indicating whether to suppress messages. Default is \code{FALSE}.}
}
\value{
The data directory.
}
\description{
This function retrieves the data directory from the environment variable SPANISH_OD_DATA_DIR.
If the environment variable is not set, it returns the temporary directory.
}
\keyword{internal}
