% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_tools.R
\name{spod_get_file_size_from_url}
\alias{spod_get_file_size_from_url}
\title{Get file size from URL}
\usage{
spod_get_file_size_from_url(x_url)
}
\arguments{
\item{x_url}{URL}
}
\value{
File size in MB
}
\description{
Get file size from URL
}
\keyword{internal}
