% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick-get.R
\name{spod_get_hmac_secret}
\alias{spod_get_hmac_secret}
\title{Get the HMAC secret from the mapas-movilidad website}
\usage{
spod_get_hmac_secret(base_url = "https://mapas-movilidad.transportes.gob.es")
}
\arguments{
\item{base_url}{The base URL of the mapas-movilidad website}
}
\value{
Character vector with the HMAC secret.
}
\description{
Get the HMAC secret from the mapas-movilidad website
}
\keyword{internal}
