% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick-get.R
\name{spod_get_hmac_secret_memoised}
\alias{spod_get_hmac_secret_memoised}
\title{Cache the HMAC secret to avoid repeated requests}
\usage{
spod_get_hmac_secret_memoised(
  base_url = "https://mapas-movilidad.transportes.gob.es"
)
}
\description{
Cache the HMAC secret to avoid repeated requests
}
\keyword{internal}
