% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{spod_get_metadata}
\alias{spod_get_metadata}
\title{Get the data dictionary}
\usage{
spod_get_metadata(data_dir = spod_get_data_dir(), quiet = FALSE)
}
\arguments{
\item{data_dir}{The directory where the data is stored. Defaults to the value returned by \code{spod_get_data_dir()}.}
}
\value{
The data dictionary.
}
\description{
This function retrieves the data dictionary for the specified data directory.
}
\examples{
# Get the data dictionary for the default data directory
if (FALSE) {
metadata = spod_get_metadata()
names(metadata)
head(metadata)
}
}
