% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-v1-data.R
\name{spod_get_od_v1}
\alias{spod_get_od_v1}
\title{Retrieve the origin-destination v1 data (2020-2021)}
\usage{
spod_get_od_v1(
  date_range = c("2020-02-14", "2020-02-15"),
  dates_list = NULL,
  date_regex = NULL,
  zones = c("distritos", "municipios"),
  data_dir = spod_get_data_dir(),
  read_fun = duckdb::tbl_file
)
}
\description{
This function retrieves the origin-destination data from the specified data directory.
}
