% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{spod_get_temp_dir}
\alias{spod_get_temp_dir}
\title{Get temporary directory for DuckDB intermediate spilling}
\usage{
spod_get_temp_dir(data_dir = spod_get_data_dir())
}
\arguments{
\item{data_dir}{The directory where the data is stored. Defaults to the value returned by \code{spod_get_data_dir()}.}
}
\value{
The path to the temp folder for DuckDB for \href{https://duckdb.org/2024/07/09/memory-management.html#intermediate-spilling}{intermediate spilling}.
}
\description{
Get the The path to the temp folder for DuckDB for \href{https://duckdb.org/2024/07/09/memory-management.html#intermediate-spilling}{intermediate spilling} in case the set memory limit and/or physical memory of the computer is too low to perform the query.
}
\keyword{internal}
