% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-utils.R
\name{spod_get_v1_meta_path}
\alias{spod_get_v1_meta_path}
\title{Get the path to the V1 metadata RDS file}
\usage{
spod_get_v1_meta_path()
}
\value{
A \code{character} string with the path to the RDS file, or \code{""} if not found.
}
\description{
Get the path to the V1 metadata RDS file
}
\keyword{internal}
