% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{spod_get_valid_dates}
\alias{spod_get_valid_dates}
\title{Get valid dates for the specified data version}
\usage{
spod_get_valid_dates(ver = 1)
}
\arguments{
\item{ver}{The version of the data to use. Defaults to 1. Can be 1 or 2.}
}
\value{
A Dates vector of valid dates for the specified data version.
}
\description{
Get valid dates for the specified data version
}
\keyword{internal}
