% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-utils.R
\name{spod_graphql_valid_dates}
\alias{spod_graphql_valid_dates}
\title{Get valid dates from the GraphQL API}
\usage{
spod_graphql_valid_dates()
}
\value{
A \code{Date} vector of dates that are valid to request data with \code{spod_quick_get_od()}.
}
\description{
Get valid dates from the GraphQL API
}
\keyword{internal}
