% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick-get.R
\name{spod_httr2_req_perform}
\alias{spod_httr2_req_perform}
\title{Internal wrappers for httr2 and sf calls to enable mocking}
\usage{
spod_httr2_req_perform(req)
}
\description{
Internal wrappers for httr2 and sf calls to enable mocking
}
\keyword{internal}
