% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-v1-data.R
\name{spod_load_zones_v1}
\alias{spod_load_zones_v1}
\title{Loads the fixed zones for v1 data}
\usage{
spod_load_zones_v1(
  type = c("distritos", "municipios"),
  data_dir = get_data_dir()
)
}
\arguments{
\item{type}{The type of zones data to retrieve ("distritos" or "municipios").}

\item{data_dir}{The directory where the data is stored.}
}
\value{
A spatial object of type \code{sf} containing the fixed zones data.
\describe{
\item{id}{\code{character}. The identifier of the zone.}
}
}
\description{
This function loads the fixed zones data for the specified data directory.
The geometry is cheked for validity and fixed if necessary. The 'ID' column is renamed to 'id'.
}
\examples{
if (FALSE) {
  zones <- spod_load_zones_v1()
}
}
