% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-utils.R
\name{spod_match_data_type_for_local_folders}
\alias{spod_match_data_type_for_local_folders}
\title{Match data types to folders}
\usage{
spod_match_data_type_for_local_folders(
  type = c("od", "origin-destination", "os", "overnight_stays", "nt", "number_of_trips",
    "rcm", "regular_commuter_mobility"),
  ver = c(1, 2)
)
}
\arguments{
\item{ver}{Integer. Can be 1 or 2. The version of the data to use. v1 spans 2020-2021, v2 covers 2022 and onwards.}
}
\value{
A \code{character} string with the folder name for the specified data type. Or \code{NULL} if the data type is not recognized.
}
\description{
Match data types to folders
}
\keyword{internal}
