% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick-get.R
\name{spod_query_od_memoised}
\alias{spod_query_od_memoised}
\title{Cache the spod_query_od_raw function to avoid repeated requests}
\usage{
spod_query_od_memoised(
  date_fmt,
  graphql_distances,
  id_origin,
  id_destination,
  min_trips,
  graphql_query
)
}
\description{
Cache the spod_query_od_raw function to avoid repeated requests
}
\keyword{internal}
