% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_helpers.R
\name{spod_read_sql}
\alias{spod_read_sql}
\title{Load an SQL query, glue it, dplyr::sql it}
\usage{
spod_read_sql(sql_file_name)
}
\arguments{
\item{sql_file_name}{The name of the SQL file to load from the package installation directory.}

\item{...}{Additional variables to pass to \code{glue::glue()} that will be evaluated in the SQL query.}
}
\value{
Text of the SQL query of class \code{sql}/\code{character}.
}
\description{
Load an SQL query from a specified file in package installation directory, glue::collapse it, glue::glue it in case of any variables that need to be replaced, and dplyr::sql it for additional safety.
}
\keyword{internal}
