% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-utils.R
\name{spod_request_length}
\alias{spod_request_length}
\title{Get the length of the request payload}
\usage{
spod_request_length(graphql_query)
}
\arguments{
\item{graphql_query}{Character. The GraphQL query.}
}
\value{
Numeric. The length of the request payload.
}
\description{
Get the length of the request payload
}
\keyword{internal}
