% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dir.R
\name{spod_set_data_dir}
\alias{spod_set_data_dir}
\title{Set the data directory}
\usage{
spod_set_data_dir(data_dir, quiet = FALSE)
}
\arguments{
\item{data_dir}{The data directory to set.}

\item{quiet}{A \code{logical} value indicating whether to suppress messages. Default is \code{FALSE}.}
}
\value{
Nothing. If quiet is \code{FALSE}, prints a message with the path and confirmation that the path exists.
}
\description{
This function sets the data directory in the environment variable SPANISH_OD_DATA_DIR, so that all other functions in the package can access the data. It also creates the directory if it doesn't exist.
}
\keyword{internal}
