% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-helpers.R
\name{spod_sql_where_dates}
\alias{spod_sql_where_dates}
\title{Generate a WHERE part of an SQL query from a sequence of dates}
\usage{
spod_sql_where_dates(dates)
}
\arguments{
\item{dates}{A Dates vector of dates to process.}
}
\value{
A character vector of the SQL query.
}
\description{
Generate a WHERE part of an SQL query from a sequence of dates
}
\keyword{internal}
