% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-tools.R
\name{spod_store_etags}
\alias{spod_store_etags}
\title{Get Etags for locally saved v1 data files and save them into a RDS file in the inst/extdata folder.}
\usage{
spod_store_etags()
}
\value{
Returns a tibble with the local path, local ETag and remote ETag.
}
\description{
Get Etags for locally saved v1 data files and save them into a RDS file in the inst/extdata folder.
}
\keyword{internal}
