% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{spod_subfolder_clean_data_cache}
\alias{spod_subfolder_clean_data_cache}
\title{Get clean data subfolder name}
\usage{
spod_subfolder_clean_data_cache(ver = 1)
}
\arguments{
\item{ver}{Integer. The version of the data. Must be 1 or 2.}
}
\value{
Character string with the subfolder name for the clean data cache.
}
\description{
Change subfolder name in the code of this function for clean data cache here to apply globally, as all functions in the package should use this function to get the clean data cache path.
}
\keyword{internal}
