% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{spod_subfolder_metadata_cache}
\alias{spod_subfolder_metadata_cache}
\title{Get metadata cache subfolder name}
\usage{
spod_subfolder_metadata_cache()
}
\value{
A \code{character} string with the subfolder name for the raw data cache.
}
\description{
Change subfolder name in the code of this function for metadata cache here to apply globally, as all functions in the package should use this function to get the metadata cache path.
}
\keyword{internal}
