% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{spod_subfolder_raw_data_cache}
\alias{spod_subfolder_raw_data_cache}
\title{Change subfolder name for raw data cache here to apply globally}
\usage{
spod_subfolder_raw_data_cache(ver = 1)
}
\arguments{
\item{ver}{Integer. The version of the data. Must be 1 or 2.}
}
\value{
Character string with the subfolder name for the raw data cache.
}
\description{
Change subfolder name for raw data cache here to apply globally
}
\keyword{internal}
