% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-utils.R
\name{spod_unique_separated_ids}
\alias{spod_unique_separated_ids}
\title{Remove duplicate values in a semicolon-separated string}
\usage{
spod_unique_separated_ids(column)
}
\arguments{
\item{column}{A \code{character} vector column in a data frame to remove duplicates from.}
}
\value{
A \code{character} vector with semicolon-separated unique IDs.
}
\description{
Remove duplicate IDs in a semicolon-separated string in a selected column in a data frame
}
\keyword{internal}
