% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_zones}
\alias{get_zones}
\title{Retrieves the zones data}
\usage{
get_zones(data_dir = get_data_dir(), type = "distritos")
}
\arguments{
\item{data_dir}{The directory where the data is stored.}

\item{type}{The type of zones data to retrieve ("distritos" or "municipios").}
}
\value{
A spatial object containing the zones data.
}
\description{
This function retrieves the zones data from the specified data directory.
It can retrieve either "distritos" or "municipios" zones data.
}
\examples{
if (FALSE) {
zones = get_zones()
}
}
