% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{spod_codebook}
\alias{spod_codebook}
\title{View codebooks for v1 and v2 open mobility data}
\usage{
spod_codebook(ver = 1)
}
\arguments{
\item{ver}{An \code{integer} or \code{numeric} value. The version of the data. Defaults to 1. Can be \code{1} for v1 (2020-2021) data and 2 for v2 (2022 onwards) data.}
}
\value{
Nothing, opens vignette if it is installed. If vignette is missing, prints a message with a link to a webpage with the codebook.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Opens relevant vignette with a codebook for v1 (2020-2021) and v2 (2022 onwards) data or provide a webpage if vignette is missing.
}
\examples{

# View codebook for v1 (2020-2021) data
spod_codebook(ver = 1)

# View codebook for v2 (2022 onwards) data
spod_codebook(ver = 2)

}
