% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_converted_data.R
\name{spod_connect_to_converted_data}
\alias{spod_connect_to_converted_data}
\title{Connect to data converted to DuckDB}
\usage{
spod_connect_to_converted_data(
  data_path,
  target_table_name = NULL,
  quiet = FALSE,
  duck_max_mem = 3,
  duck_max_threads = parallelly::availableCores() - 1
)
}
\arguments{
\item{data_path}{a path to the \code{DuckDB} database file with '.duckdb' extension, or a path to the folder with \code{parquet} files. Eigher one should have been created with the \code{spod_convert_for_analysis()} function.}

\item{target_table_name}{table name inside the database.}

\item{duck_max_mem}{The maximum memory to use in GB. A conservative default is 3 GB, which should be enough for resaving the data to DuckDB form a folder of CSV.gz files while being small enough to fit in memory of most even old computers. For data analysis using the already converted data (in DuckDB or Parquet format) or with the raw CSV.gz data, it is recommended to increase it according to available resources.}

\item{duck_max_threads}{The maximum number of threads to use. Defaults to the number of available cores minus 1.}
}
\value{
a DuckDB table connection object.
}
\description{
This function allows the user to quickly connect to the data converted to DuckDB with the \code{spod_convert_to_duckdb()} function. This function is a simplificaiton of the connection process. It uses
}
