% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{spod_get}
\alias{spod_get}
\title{Retrieves the origin-destination data}
\usage{
spod_get(
  data_dir = spod_get_data_dir(),
  subdir = "estudios_basicos/por-distritos/viajes/ficheros-diarios",
  date_regex = "2024030[1-2]",
  read_fun = duckdb::tbl_file
)
}
\arguments{
\item{data_dir}{The directory where the data is stored.}

\item{subdir}{The subdirectory where the data is stored.}

\item{date_regex}{The regular expression to match the date of the data to download.}

\item{read_fun}{The function to read the data. Defaults to \code{duckdb::tbl_file}.}
}
\value{
The local path of the downloaded file (\code{download_od}), or a data frame with the origin-destination data (\code{spod_get}).
}
\description{
This function downloads data from URLs such as
https://movilidad-opendata.mitma.es/estudios_basicos/por-distritos/viajes/ficheros-diarios/2024-03/20240301_Viajes_distritos.csv.gz
if the file does not exist in the data directory.
}
\examples{
# Download the origin-destination data for the first two days of March 2024
if (FALSE) {
od_20240301_20240302 = spod_get(date_regex = "2024-03-0[1-2]")
}
}
