% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dir.R
\name{spod_get_data_dir}
\alias{spod_get_data_dir}
\title{Get the data directory}
\usage{
spod_get_data_dir(quiet = FALSE)
}
\arguments{
\item{quiet}{A \code{logical} value indicating whether to suppress messages. Default is \code{FALSE}.}
}
\value{
A \code{character} vector of length 1 containing the path to the data directory where the package will download and convert the data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function retrieves the data directory from the environment variable SPANISH_OD_DATA_DIR (and previously set by \code{\link[=spod_set_data_dir]{spod_set_data_dir()}}).
If the environment variable is not set, it returns the temporary directory.
}
\examples{
spod_set_data_dir(tempdir())
spod_get_data_dir()

}
