% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{spod_get_valid_dates}
\alias{spod_get_valid_dates}
\title{Get valid dates for the specified data version}
\usage{
spod_get_valid_dates(ver = NULL)
}
\arguments{
\item{ver}{Integer. Can be 1 or 2. The version of the data to use. v1 spans 2020-2021, v2 covers 2022 and onwards.}
}
\value{
A vector of type \code{Date} with all possible valid dates for the specified data version (v1 for 2020-2021 and v2 for 2020 onwards).
}
\description{
Get valid dates for the specified data version
}
