% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-v1-data.R
\name{spod_get_zones_v1}
\alias{spod_get_zones_v1}
\title{Retrieves the zones for v1 data}
\usage{
spod_get_zones_v1(
  type = c("distritos", "municipios"),
  data_dir = get_data_dir()
)
}
\arguments{
\item{type}{The type of zones data to retrieve ("distritos" or "municipios").}

\item{data_dir}{The directory where the data is stored.}
}
\value{
A spatial object containing the zones data.
}
\description{
This function retrieves the zones data from the specified data directory.
It can retrieve either "distritos" or "municipios" zones data.
}
\examples{
if (FALSE) {
  zones <- get_zones()
}
}
