% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-utils.R
\name{spod_match_data_type}
\alias{spod_match_data_type}
\title{Match data types for normalisation}
\usage{
spod_match_data_type(
  type = c("od", "origin-destination", "viajes", "os", "overnight_stays",
    "pernoctaciones", "nt", "number_of_trips", "personas")
)
}
\arguments{
\item{type}{The type of data to match. Can be "od", "origin-destination", "os", "overnight_stays", or "nt", "number_of_trips".}
}
\description{
Match data types for normalisation
}
\keyword{internal}
