% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{spod_subfolder_raw_data_cache}
\alias{spod_subfolder_raw_data_cache}
\title{Get raw data cache subfolder name}
\usage{
spod_subfolder_raw_data_cache(ver = 1)
}
\arguments{
\item{ver}{Integer. Can be 1 or 2. The version of the data to use. v1 spans 2020-2021, v2 covers 2022 and onwards. See more details in codebooks with \code{\link[=spod_codebook]{spod_codebook()}}.}
}
\value{
A \code{character} string with the subfolder name for the raw data cache.
}
\description{
Change subfolder name in the code of this function for raw data cache here to apply globally, as all functions in the package should use this function to get the raw data cache path.
}
\keyword{internal}
